Refactor eval and add interpreter runtime tests
