+++ src/interpreter/eval.rs
@@ -1,5 +1,9 @@
 use super::*;
 
+mod access;
+mod literals;
+mod modules;
+
 /// StringToBigInt per spec §7.1.14 — handles empty string, hex, octal, binary.
 fn string_to_bigint_for_comparison(s: &str) -> Option<num_bigint::BigInt> {
     let trimmed = s.trim();
@@ -1280,640 +1284,7 @@ impl Interpreter {
         }
     }
 
-    /// Evaluate an OptionalChain expression and return (value, this_context).
-    /// Used when the optional chain result feeds into a call or nested chain.
-    fn eval_optional_chain_with_ref(
-        &mut self,
-        base: &Expression,
-        chain: &Expression,
-        env: &EnvRef,
-    ) -> Result<(JsValue, JsValue), Completion> {
-        let (base_val, base_this) = self.eval_oc_base(base, chain, env)?;
-        if matches!(base_val, JsValue::Null | JsValue::Undefined) {
-            return Ok((JsValue::Undefined, JsValue::Undefined));
-        }
-        self.eval_oc_tail_with_this_ctx(&base_val, &base_this, chain, env)
-    }
-
-    /// Evaluate the base expression of an OptionalChain, returning (value, this).
-    fn eval_oc_base(
-        &mut self,
-        base: &Expression,
-        chain: &Expression,
-        env: &EnvRef,
-    ) -> Result<(JsValue, JsValue), Completion> {
-        match base {
-            Expression::Member(obj_expr, member_prop) => {
-                if matches!(obj_expr.as_ref(), Expression::Super) {
-                    // §13.3.7.1: super property in optional chain — use HomeObject.__proto__
-                    let this_val = env.borrow().get("this").unwrap_or(JsValue::Undefined);
-                    let key = match member_prop {
-                        MemberProperty::Dot(name) => name.clone(),
-                        MemberProperty::Computed(expr) => {
-                            let v = match self.eval_expr(expr, env) {
-                                Completion::Normal(v) => v,
-                                other => return Err(other),
-                            };
-                            match self.to_property_key(&v) {
-                                Ok(s) => s,
-                                Err(e) => return Err(Completion::Throw(e)),
-                            }
-                        }
-                        MemberProperty::Private(name) => {
-                            let branded = self.resolve_private_name(name, env);
-                            if let JsValue::Object(ref o) = this_val
-                                && let Some(obj) = self.get_object(o.id)
-                            {
-                                let elem = obj.borrow().private_fields.get(&branded).cloned();
-                                match elem {
-                                    Some(PrivateElement::Field(v))
-                                    | Some(PrivateElement::Method(v)) => {
-                                        return Ok((v, this_val));
-                                    }
-                                    Some(PrivateElement::Accessor { get, .. }) => {
-                                        if let Some(getter) = get {
-                                            match self.call_function(&getter, &this_val, &[]) {
-                                                Completion::Normal(v) => return Ok((v, this_val)),
-                                                other => return Err(other),
-                                            }
-                                        }
-                                        return Err(Completion::Throw(self.create_type_error(
-                                            &format!("Cannot read private member #{name}"),
-                                        )));
-                                    }
-                                    None => {
-                                        return Err(Completion::Throw(self.create_type_error(
-                                            &format!("Cannot read private member #{name}"),
-                                        )));
-                                    }
-                                }
-                            }
-                            return Ok((JsValue::Undefined, this_val));
-                        }
-                    };
-                    let super_base_id = self.get_super_base_id(env);
-                    match super_base_id {
-                        Some(base_id) => {
-                            let val = match self.get_object_property(base_id, &key, &this_val) {
-                                Completion::Normal(v) => v,
-                                other => return Err(other),
-                            };
-                            Ok((val, this_val))
-                        }
-                        None => Err(Completion::Throw(self.create_type_error(&format!(
-                            "Cannot read properties of null (reading '{key}')"
-                        )))),
-                    }
-                } else {
-                    let obj_val = match self.eval_expr(obj_expr, env) {
-                        Completion::Normal(v) => v,
-                        other => return Err(other),
-                    };
-                    let key = match member_prop {
-                        MemberProperty::Dot(name) => name.clone(),
-                        MemberProperty::Computed(expr) => {
-                            let v = match self.eval_expr(expr, env) {
-                                Completion::Normal(v) => v,
-                                other => return Err(other),
-                            };
-                            match self.to_property_key(&v) {
-                                Ok(s) => s,
-                                Err(e) => return Err(Completion::Throw(e)),
-                            }
-                        }
-                        MemberProperty::Private(name) => {
-                            let branded = self.resolve_private_name(name, env);
-                            if let JsValue::Object(ref o) = obj_val
-                                && let Some(obj) = self.get_object(o.id)
-                            {
-                                let elem = obj.borrow().private_fields.get(&branded).cloned();
-                                match elem {
-                                    Some(PrivateElement::Field(v))
-                                    | Some(PrivateElement::Method(v)) => {
-                                        if matches!(v, JsValue::Null | JsValue::Undefined) {
-                                            return Ok((JsValue::Undefined, JsValue::Undefined));
-                                        }
-                                        return self.eval_oc_tail_with_this(&v, chain, env);
-                                    }
-                                    Some(PrivateElement::Accessor { get, .. }) => {
-                                        if let Some(getter) = get {
-                                            let v = match self.call_function(&getter, &obj_val, &[])
-                                            {
-                                                Completion::Normal(v) => v,
-                                                other => return Err(other),
-                                            };
-                                            if matches!(v, JsValue::Null | JsValue::Undefined) {
-                                                return Ok((
-                                                    JsValue::Undefined,
-                                                    JsValue::Undefined,
-                                                ));
-                                            }
-                                            return self.eval_oc_tail_with_this(&v, chain, env);
-                                        }
-                                        return Ok((JsValue::Undefined, JsValue::Undefined));
-                                    }
-                                    None => {
-                                        return Err(Completion::Throw(self.create_type_error(
-                                            &format!("Cannot read private member #{name}"),
-                                        )));
-                                    }
-                                }
-                            } else {
-                                return Ok((JsValue::Undefined, JsValue::Undefined));
-                            }
-                        }
-                    };
-                    let prop_val = match self.access_property_on_value(&obj_val, &key) {
-                        Completion::Normal(v) => v,
-                        other => return Err(other),
-                    };
-                    Ok((prop_val, obj_val))
-                }
-            }
-            Expression::OptionalChain(inner_base, inner_chain) => {
-                // Nested optional chain: preserve this context from inner chain
-                self.eval_optional_chain_with_ref(inner_base, inner_chain, env)
-            }
-            _ => {
-                let val = match self.eval_expr(base, env) {
-                    Completion::Normal(v) => v,
-                    other => return Err(other),
-                };
-                Ok((val, JsValue::Undefined))
-            }
-        }
-    }
-
-    /// Evaluate optional chain tail with a known `this` from the base member access.
-    /// This is used when the optional chain base is `obj.method?.()` so that
-    /// the call uses `obj` as `this`.
-    fn eval_optional_chain_tail_with_base_this(
-        &mut self,
-        base_val: &JsValue,
-        base_this: &JsValue,
-        prop: &Expression,
-        env: &EnvRef,
-    ) -> Completion {
-        match self.eval_oc_tail_with_this_ctx(base_val, base_this, prop, env) {
-            Ok((v, _)) => Completion::Normal(v),
-            Err(c) => c,
-        }
-    }
-
-    /// Evaluate optional chain tail, returning (value, this_for_call).
-    fn eval_oc_tail_with_this(
-        &mut self,
-        base_val: &JsValue,
-        prop: &Expression,
-        env: &EnvRef,
-    ) -> Result<(JsValue, JsValue), Completion> {
-        self.eval_oc_tail_with_this_ctx(base_val, &JsValue::Undefined, prop, env)
-    }
-
-    /// Core optional chain tail evaluator with explicit this context.
-    /// `chain_this` is the `this` value to use for `?.()` direct calls
-    /// (from `obj.method?.()` where chain_this = obj).
-    fn eval_oc_tail_with_this_ctx(
-        &mut self,
-        base_val: &JsValue,
-        chain_this: &JsValue,
-        prop: &Expression,
-        env: &EnvRef,
-    ) -> Result<(JsValue, JsValue), Completion> {
-        match prop {
-            Expression::Identifier(name) => {
-                if name.is_empty() {
-                    // x?.() — direct call placeholder, base_val IS the value
-                    // chain_this is the object for `obj.method?.()` calls
-                    Ok((base_val.clone(), chain_this.clone()))
-                } else {
-                    let val = match self.access_property_on_value(base_val, name) {
-                        Completion::Normal(v) => v,
-                        Completion::Throw(e) => return Err(Completion::Throw(e)),
-                        other => return Err(other),
-                    };
-                    Ok((val, base_val.clone()))
-                }
-            }
-            Expression::Call(callee, args) => {
-                let (func_val, this_val) =
-                    self.eval_oc_tail_with_this_ctx(base_val, chain_this, callee, env)?;
-                let evaluated_args = match self.eval_spread_args(args, env) {
-                    Ok(v) => v,
-                    Err(e) => return Err(Completion::Throw(e)),
-                };
-                match self.call_function(&func_val, &this_val, &evaluated_args) {
-                    Completion::Normal(v) => Ok((v, JsValue::Undefined)),
-                    other => Err(other),
-                }
-            }
-            Expression::Member(inner, mp) => {
-                let (inner_val, _) =
-                    self.eval_oc_tail_with_this_ctx(base_val, chain_this, inner, env)?;
-                // Non-optional member access within optional chain: null/undefined throws
-                if matches!(&inner_val, JsValue::Null | JsValue::Undefined) {
-                    let key_str = match mp {
-                        MemberProperty::Dot(name) => name.clone(),
-                        MemberProperty::Computed(_) => "property".to_string(),
-                        MemberProperty::Private(name) => format!("#{name}"),
-                    };
-                    return Err(Completion::Throw(self.create_type_error(&format!(
-                        "Cannot read properties of {} (reading '{key_str}')",
-                        if matches!(&inner_val, JsValue::Null) {
-                            "null"
-                        } else {
-                            "undefined"
-                        }
-                    ))));
-                }
-                match mp {
-                    MemberProperty::Dot(name) => {
-                        let val = match self.access_property_on_value(&inner_val, name) {
-                            Completion::Normal(v) => v,
-                            other => return Err(other),
-                        };
-                        Ok((val, inner_val))
-                    }
-                    MemberProperty::Computed(expr) => {
-                        let key_val = match self.eval_expr(expr, env) {
-                            Completion::Normal(v) => v,
-                            other => return Err(other),
-                        };
-                        let key = match self.to_property_key(&key_val) {
-                            Ok(s) => s,
-                            Err(e) => return Err(Completion::Throw(e)),
-                        };
-                        let val = match self.access_property_on_value(&inner_val, &key) {
-                            Completion::Normal(v) => v,
-                            other => return Err(other),
-                        };
-                        Ok((val, inner_val))
-                    }
-                    MemberProperty::Private(name) => {
-                        let branded = self.resolve_private_name(name, env);
-                        if let JsValue::Object(o) = &inner_val
-                            && let Some(obj) = self.get_object(o.id)
-                        {
-                            let elem = obj.borrow().private_fields.get(&branded).cloned();
-                            match elem {
-                                Some(PrivateElement::Field(v))
-                                | Some(PrivateElement::Method(v)) => {
-                                    Ok((v, inner_val))
-                                }
-                                Some(PrivateElement::Accessor { get, .. }) => {
-                                    if let Some(getter) = get {
-                                        match self.call_function(&getter, &inner_val, &[]) {
-                                            Completion::Normal(v) => Ok((v, inner_val)),
-                                            other => Err(other),
-                                        }
-                                    } else {
-                                        Err(Completion::Throw(self.create_type_error(&format!(
-                                            "Cannot read private member #{name} which has no getter"
-                                        ))))
-                                    }
-                                }
-                                None => Err(Completion::Throw(self.create_type_error(&format!(
-                                    "Cannot read private member #{name} from an object whose class did not declare it"
-                                )))),
-                            }
-                        } else {
-                            Ok((JsValue::Undefined, inner_val))
-                        }
-                    }
-                }
-            }
-            other => {
-                // Computed property access (e.g., x?.[expr])
-                let key_val = match self.eval_expr(other, env) {
-                    Completion::Normal(v) => v,
-                    other => return Err(other),
-                };
-                let key = match self.to_property_key(&key_val) {
-                    Ok(s) => s,
-                    Err(e) => return Err(Completion::Throw(e)),
-                };
-                let val = match self.access_property_on_value(base_val, &key) {
-                    Completion::Normal(v) => v,
-                    other => return Err(other),
-                };
-                Ok((val, base_val.clone()))
-            }
-        }
-    }
-
-    /// Handle `delete obj?.prop` and `delete obj?.['prop']` etc.
-    fn eval_delete_optional_chain(
-        &mut self,
-        base: &Expression,
-        chain: &Expression,
-        env: &EnvRef,
-    ) -> Completion {
-        // Evaluate the base of the optional chain
-        let (base_val, _base_this) = match self.eval_oc_base(base, chain, env) {
-            Ok(v) => v,
-            Err(c) => return c,
-        };
-        // If base is null/undefined, short-circuit to true
-        if matches!(base_val, JsValue::Null | JsValue::Undefined) {
-            return Completion::Normal(JsValue::Boolean(true));
-        }
-        // Walk the chain to find the object and key to delete from
-        self.eval_delete_oc_tail(&base_val, chain, env)
-    }
-
-    fn eval_delete_oc_tail(
-        &mut self,
-        base_val: &JsValue,
-        chain: &Expression,
-        env: &EnvRef,
-    ) -> Completion {
-        match chain {
-            Expression::Identifier(name) if !name.is_empty() => {
-                // delete obj?.prop → delete obj.prop
-                self.eval_delete_on_object(base_val, name, env)
-            }
-            Expression::Member(inner, mp) => {
-                // Evaluate inner to get the object, then delete the last property
-                let (inner_val, _) = match self.eval_oc_tail_with_this_ctx(
-                    base_val,
-                    &JsValue::Undefined,
-                    inner,
-                    env,
-                ) {
-                    Ok(v) => v,
-                    Err(c) => return c,
-                };
-                // Check null/undefined for chained access
-                if matches!(&inner_val, JsValue::Null | JsValue::Undefined) {
-                    return Completion::Normal(JsValue::Boolean(true));
-                }
-                let key = match mp {
-                    MemberProperty::Dot(name) => name.clone(),
-                    MemberProperty::Computed(expr) => {
-                        let v = match self.eval_expr(expr, env) {
-                            Completion::Normal(v) => v,
-                            other => return other,
-                        };
-                        match self.to_property_key(&v) {
-                            Ok(s) => s,
-                            Err(e) => return Completion::Throw(e),
-                        }
-                    }
-                    MemberProperty::Private(_) => {
-                        return Completion::Throw(
-                            self.create_type_error("Private fields cannot be deleted"),
-                        );
-                    }
-                };
-                self.eval_delete_on_object(&inner_val, &key, env)
-            }
-            Expression::Call(callee, args) => {
-                // delete obj?.method() — evaluate the call for side effects, return true
-                let (func_val, this_val) = match self.eval_oc_tail_with_this_ctx(
-                    base_val,
-                    &JsValue::Undefined,
-                    callee,
-                    env,
-                ) {
-                    Ok(v) => v,
-                    Err(c) => return c,
-                };
-                let evaluated_args = match self.eval_spread_args(args, env) {
-                    Ok(v) => v,
-                    Err(e) => return Completion::Throw(e),
-                };
-                match self.call_function(&func_val, &this_val, &evaluated_args) {
-                    Completion::Normal(_) => Completion::Normal(JsValue::Boolean(true)),
-                    other => other,
-                }
-            }
-            _ => {
-                // Fallback: evaluate the chain for side effects, return true
-                match self.eval_optional_chain_tail_with_base_this(
-                    base_val,
-                    &JsValue::Undefined,
-                    chain,
-                    env,
-                ) {
-                    Completion::Normal(_) => Completion::Normal(JsValue::Boolean(true)),
-                    other => other,
-                }
-            }
-        }
-    }
-
-    fn eval_delete_on_object(&mut self, obj_val: &JsValue, key: &str, env: &EnvRef) -> Completion {
-        let obj_val = if !matches!(obj_val, JsValue::Object(_)) {
-            match self.to_object(obj_val) {
-                Completion::Normal(v) => v,
-                Completion::Throw(e) => return Completion::Throw(e),
-                _ => return Completion::Normal(JsValue::Boolean(true)),
-            }
-        } else {
-            obj_val.clone()
-        };
-        if let JsValue::Object(ref o) = obj_val
-            && let Some(obj) = self.get_object(o.id)
-        {
-            if obj.borrow().is_proxy() || obj.borrow().proxy_revoked {
-                match self.proxy_delete_property(o.id, key) {
-                    Ok(false) => {
-                        if env.borrow().strict {
-                            return Completion::Throw(self.create_type_error(&format!(
-                                "Cannot delete property '{key}' of object"
-                            )));
-                        }
-                        return Completion::Normal(JsValue::Boolean(false));
-                    }
-                    Ok(result) => return Completion::Normal(JsValue::Boolean(result)),
-                    Err(e) => return Completion::Throw(e),
-                }
-            }
-            let is_strict = env.borrow().strict;
-            let mut obj_mut = obj.borrow_mut();
-            if let Some(desc) = obj_mut.properties.get(key)
-                && desc.configurable == Some(false)
-            {
-                if is_strict {
-                    drop(obj_mut);
-                    return Completion::Throw(
-                        self.create_type_error(&format!(
-                            "Cannot delete property '{key}' of object"
-                        )),
-                    );
-                }
-                return Completion::Normal(JsValue::Boolean(false));
-            }
-            obj_mut.properties.remove(key);
-            obj_mut.property_order.retain(|k| k != key);
-            if let Some(ref mut map) = obj_mut.parameter_map {
-                map.remove(key);
-            }
-            if let Ok(idx) = key.parse::<usize>()
-                && let Some(ref mut elems) = obj_mut.array_elements
-                && idx < elems.len()
-            {
-                elems[idx] = JsValue::Undefined;
-            }
-        }
-        Completion::Normal(JsValue::Boolean(true))
-    }
-
-    fn get_template_object(&mut self, tmpl: &TemplateLiteral) -> JsValue {
-        let cache_key = tmpl.id;
-        if let Some(&obj_id) = self.realm().template_cache.get(&cache_key)
-            && self.get_object(obj_id).is_some()
-        {
-            return JsValue::Object(crate::types::JsObject { id: obj_id });
-        }
-
-        let cooked_vals: Vec<JsValue> = tmpl
-            .quasis
-            .iter()
-            .map(|q| match q {
-                Some(cu) => JsValue::String(JsString {
-                    code_units: cu.clone(),
-                }),
-                None => JsValue::Undefined,
-            })
-            .collect();
-        let raw_vals: Vec<JsValue> = tmpl
-            .raw_quasis
-            .iter()
-            .map(|s| JsValue::String(JsString::from_str(s)))
-            .collect();
-
-        let raw_arr = self.create_frozen_template_array(raw_vals);
-        let template_arr = self.create_frozen_template_array(cooked_vals);
-
-        if let JsValue::Object(o) = &template_arr
-            && let Some(obj) = self.get_object(o.id)
-        {
-            obj.borrow_mut().insert_property(
-                "raw".to_string(),
-                PropertyDescriptor::data(raw_arr, false, false, false),
-            );
-        }
-
-        if let JsValue::Object(o) = &template_arr {
-            self.realm_mut().template_cache.insert(cache_key, o.id);
-        }
-
-        template_arr
-    }
-
-    fn create_frozen_template_array(&mut self, values: Vec<JsValue>) -> JsValue {
-        let len = values.len();
-        let mut obj_data = JsObjectData::new();
-        obj_data.prototype = self
-            .realm()
-            .array_prototype
-            .clone()
-            .or(self.realm().object_prototype.clone());
-        obj_data.class_name = "Array".to_string();
-        for (i, v) in values.iter().enumerate() {
-            obj_data.insert_property(
-                i.to_string(),
-                PropertyDescriptor::data(v.clone(), false, true, false),
-            );
-        }
-        obj_data.insert_property(
-            "length".to_string(),
-            PropertyDescriptor::data(JsValue::Number(len as f64), false, false, false),
-        );
-        obj_data.array_elements = Some(values);
-        obj_data.extensible = false;
-        let obj = Rc::new(RefCell::new(obj_data));
-        let id = self.allocate_object_slot(obj);
-        JsValue::Object(crate::types::JsObject { id })
-    }
-
-    fn eval_literal(&mut self, lit: &Literal) -> JsValue {
-        match lit {
-            Literal::Null => JsValue::Null,
-            Literal::Boolean(b) => JsValue::Boolean(*b),
-            Literal::Number(n) => JsValue::Number(*n),
-            Literal::String(s) => {
-                let code_units =
-                    crate::interpreter::builtins::regexp::pua_code_units_to_surrogates(s);
-                JsValue::String(JsString { code_units })
-            }
-            Literal::BigInt(s) => {
-                use num_bigint::BigInt;
-                let value = if let Some(hex) = s.strip_prefix("0x").or_else(|| s.strip_prefix("0X"))
-                {
-                    BigInt::parse_bytes(hex.as_bytes(), 16).unwrap_or_default()
-                } else if let Some(oct) = s.strip_prefix("0o").or_else(|| s.strip_prefix("0O")) {
-                    BigInt::parse_bytes(oct.as_bytes(), 8).unwrap_or_default()
-                } else if let Some(bin) = s.strip_prefix("0b").or_else(|| s.strip_prefix("0B")) {
-                    BigInt::parse_bytes(bin.as_bytes(), 2).unwrap_or_default()
-                } else {
-                    s.parse::<BigInt>().unwrap_or_default()
-                };
-                JsValue::BigInt(JsBigInt { value })
-            }
-            Literal::RegExp(pattern, flags) => {
-                let mut obj = JsObjectData::new();
-                obj.prototype = self
-                    .realm()
-                    .regexp_prototype
-                    .clone()
-                    .or(self.realm().object_prototype.clone());
-                obj.class_name = "RegExp".to_string();
-                let source_js = if pattern.is_empty() {
-                    JsString::from_str("(?:)")
-                } else {
-                    crate::interpreter::builtins::regexp::regex_output_to_js_string(pattern)
-                };
-                obj.regexp_original_source = Some(source_js);
-                obj.regexp_original_flags = Some(JsString::from_str(flags));
-                obj.insert_property(
-                    "lastIndex".to_string(),
-                    PropertyDescriptor::data(JsValue::Number(0.0), true, false, false),
-                );
-                let rc = Rc::new(RefCell::new(obj));
-                let id = self.allocate_object_slot(rc);
-                JsValue::Object(crate::types::JsObject { id })
-            }
-        }
-    }
-
     // §7.1.14 ToPropertyKey
-    #[allow(clippy::wrong_self_convention)]
-    pub(crate) fn to_property_key(&mut self, val: &JsValue) -> Result<String, JsValue> {
-        match val {
-            JsValue::Symbol(s) => Ok(s.to_property_key()),
-            JsValue::Object(_) => {
-                let prim = self.to_primitive(val, "string")?;
-                if let JsValue::Symbol(s) = &prim {
-                    return Ok(s.to_property_key());
-                }
-                self.to_string_value(&prim)
-            }
-            _ => self.to_string_value(val),
-        }
-    }
-
-    pub(crate) fn create_regexp(&mut self, pattern: &str, flags: &str) -> JsValue {
-        let mut obj = JsObjectData::new();
-        obj.prototype = self
-            .realm()
-            .regexp_prototype
-            .clone()
-            .or(self.realm().object_prototype.clone());
-        obj.class_name = "RegExp".to_string();
-        let source_str = if pattern.is_empty() { "(?:)" } else { pattern };
-        obj.regexp_original_source = Some(JsString::from_str(source_str));
-        obj.regexp_original_flags = Some(JsString::from_str(flags));
-        obj.insert_property(
-            "lastIndex".to_string(),
-            PropertyDescriptor::data(JsValue::Number(0.0), true, false, false),
-        );
-        let rc = Rc::new(RefCell::new(obj));
-        let id = self.allocate_object_slot(rc);
-        JsValue::Object(crate::types::JsObject { id })
-    }
-
     fn eval_unary(&mut self, op: UnaryOp, val: &JsValue) -> Completion {
         match op {
             UnaryOp::Minus => {
@@ -14366,588 +13737,6 @@ impl Interpreter {
         Completion::Normal(JsValue::Boolean(false))
     }
 
-    /// Resolve a module export value by following re-export chains recursively.
-    /// Used by namespace [[Get]] to dynamically resolve live bindings.
-    fn resolve_module_export_value(
-        &mut self,
-        binding_name: &str,
-        env: &crate::interpreter::types::EnvRef,
-        module_path: Option<&std::path::Path>,
-        original_key: &str,
-    ) -> Result<JsValue, JsValue> {
-        self.resolve_module_export_value_inner(
-            binding_name,
-            env,
-            module_path,
-            original_key,
-            &mut std::collections::HashSet::new(),
-        )
-    }
-
-    fn resolve_module_export_value_inner(
-        &mut self,
-        binding_name: &str,
-        env: &crate::interpreter::types::EnvRef,
-        module_path: Option<&std::path::Path>,
-        original_key: &str,
-        visited: &mut std::collections::HashSet<(std::path::PathBuf, String)>,
-    ) -> Result<JsValue, JsValue> {
-        if let Some(mp) = module_path {
-            let key = (mp.to_path_buf(), binding_name.to_string());
-            if visited.contains(&key) {
-                return Err(self.create_reference_error(&format!(
-                    "Cannot access '{}' before initialization",
-                    original_key
-                )));
-            }
-            visited.insert(key);
-        }
-
-        // Handle *ns: bindings (namespace re-export)
-        if binding_name.starts_with("*ns:") {
-            if let Some(val) = env.borrow().get(original_key) {
-                return Ok(val);
-            }
-            if let Some(mp) = module_path
-                && let Some(module) = self.module_registry.get(mp)
-                && let Some(val) = module.borrow().exports.get(original_key)
-            {
-                return Ok(val.clone());
-            }
-            return Ok(JsValue::Undefined);
-        }
-
-        // Handle *reexport:source:name — follow the chain recursively
-        if let Some(rest) = binding_name.strip_prefix("*reexport:") {
-            if let Some(colon_idx) = rest.rfind(':') {
-                let source = &rest[..colon_idx];
-                let export_name = &rest[colon_idx + 1..];
-                if let Some(mp) = module_path
-                    && let Ok(resolved) = self.resolve_module_specifier(source, Some(mp))
-                    && let Ok(source_mod) = self.load_module(&resolved)
-                {
-                    let source_ref = source_mod.borrow();
-                    let source_env = source_ref.env.clone();
-                    let source_path = source_ref.path.clone();
-                    // Look up what this export resolves to in the source module
-                    if let Some(next_binding) = source_ref.export_bindings.get(export_name) {
-                        let next_binding = next_binding.clone();
-                        drop(source_ref);
-                        return self.resolve_module_export_value_inner(
-                            &next_binding,
-                            &source_env,
-                            Some(&source_path),
-                            original_key,
-                            visited,
-                        );
-                    }
-                    drop(source_ref);
-                    // No binding info — try direct env lookup
-                    if source_env.borrow().is_in_tdz(export_name) {
-                        return Err(self.create_reference_error(&format!(
-                            "Cannot access '{}' before initialization",
-                            original_key
-                        )));
-                    }
-                    if let Some(val) = source_env.borrow().get(export_name) {
-                        return Ok(val);
-                    }
-                }
-            }
-            return Ok(JsValue::Undefined);
-        }
-
-        // Local binding: look up in the provided environment
-        if env.borrow().is_in_tdz(binding_name) {
-            return Err(self.create_reference_error(&format!(
-                "Cannot access '{}' before initialization",
-                original_key
-            )));
-        }
-        if let Some(val) = env.borrow().get(binding_name) {
-            return Ok(val);
-        }
-
-        Ok(JsValue::Undefined)
-    }
-
-    /// Check if accessing `key` on a module namespace object would hit TDZ.
-    /// Returns Err(ReferenceError) if the binding is uninitialized.
-    /// Returns Ok(()) if the key is safe to access or the object is not a namespace.
-    pub(crate) fn check_namespace_tdz(&mut self, obj_id: u64, key: &str) -> Result<(), JsValue> {
-        if key.starts_with("Symbol(") {
-            return Ok(());
-        }
-        let ns_data = if let Some(obj) = self.get_object(obj_id) {
-            obj.borrow().module_namespace.clone()
-        } else {
-            None
-        };
-        if let Some(ns_data) = ns_data {
-            // Deferred namespaces: trigger evaluation on non-symbol-like key access
-            if ns_data.deferred && !Self::is_symbol_like_namespace_key(key, true) {
-                self.ensure_deferred_namespace_evaluation(obj_id)?;
-            }
-            if let Some(binding_name) = ns_data.export_to_binding.get(key) {
-                if binding_name.starts_with("*ns:") {
-                    return Ok(());
-                }
-                if let Some(rest) = binding_name.strip_prefix("*reexport:") {
-                    // Check TDZ in source module's environment
-                    if let Some(colon_idx) = rest.rfind(':') {
-                        let source = &rest[..colon_idx];
-                        let export_name = &rest[colon_idx + 1..];
-                        if let Some(ref module_path) = ns_data.module_path
-                            && let Ok(resolved) =
-                                self.resolve_module_specifier(source, Some(module_path))
-                            && let Ok(source_mod) = self.load_module(&resolved)
-                        {
-                            let source_ref = source_mod.borrow();
-                            if let Some(binding) = source_ref.export_bindings.get(export_name) {
-                                if source_ref.env.borrow().is_in_tdz(binding) {
-                                    return Err(self.create_reference_error(&format!(
-                                        "Cannot access '{key}' before initialization"
-                                    )));
-                                }
-                            } else if source_ref.env.borrow().is_in_tdz(export_name) {
-                                return Err(self.create_reference_error(&format!(
-                                    "Cannot access '{key}' before initialization"
-                                )));
-                            }
-                        }
-                    }
-                    return Ok(());
-                }
-                if ns_data.env.borrow().is_in_tdz(binding_name) {
-                    return Err(self.create_reference_error(&format!(
-                        "Cannot access '{key}' before initialization"
-                    )));
-                }
-            }
-        }
-        Ok(())
-    }
-
-    /// IsSymbolLikeNamespaceKey(P, O): true if P is a Symbol, or deferred + "then"
-    pub fn is_symbol_like_namespace_key(key: &str, deferred: bool) -> bool {
-        key.starts_with("Symbol(") || (deferred && key == "then")
-    }
-
-    /// Trigger evaluation of a deferred module namespace when a non-symbol-like key is accessed.
-    pub(crate) fn ensure_deferred_namespace_evaluation(
-        &mut self,
-        obj_id: u64,
-    ) -> Result<(), JsValue> {
-        let (deferred, module_path) = if let Some(obj) = self.get_object(obj_id) {
-            let b = obj.borrow();
-            if let Some(ref ns) = b.module_namespace {
-                (ns.deferred, ns.module_path.clone())
-            } else {
-                return Ok(());
-            }
-        } else {
-            return Ok(());
-        };
-
-        if !deferred {
-            return Ok(());
-        }
-
-        let module_path = match module_path {
-            Some(p) => p,
-            None => return Ok(()),
-        };
-
-        let module = match self.module_registry.get(&module_path).cloned() {
-            Some(m) => m,
-            None => return Ok(()),
-        };
-
-        if module.borrow().evaluated {
-            // Already evaluated — just clear deferred flag
-            if let Some(ref err) = module.borrow().error {
-                return Err(err.clone());
-            }
-            if let Some(obj) = self.get_object(obj_id)
-                && let Some(ref mut ns) = obj.borrow_mut().module_namespace
-            {
-                ns.deferred = false;
-            }
-            return Ok(());
-        }
-
-        // Check ReadyForSyncExecution
-        let mut seen = std::collections::HashSet::new();
-        if !self.ready_for_sync_execution(&module_path, &mut seen) {
-            return Err(self.create_type_error(
-                "Cannot synchronously evaluate a module with top-level await or that is currently being evaluated",
-            ));
-        }
-
-        // Save and set current_module_path for evaluation
-        let prev_path = self.current_module_path.take();
-        self.current_module_path = Some(module_path.clone());
-        let mut stack = vec![];
-        let result = self.inner_module_evaluation(&module_path, &mut stack, 0);
-        self.current_module_path = prev_path;
-
-        match result {
-            Ok(_) => {
-                if let Some(obj) = self.get_object(obj_id)
-                    && let Some(ref mut ns) = obj.borrow_mut().module_namespace
-                {
-                    ns.deferred = false;
-                }
-                Ok(())
-            }
-            Err(ref e) => {
-                // Per spec §16.2.1.5.3 step 9: mark all modules on stack as evaluated with error
-                for m_path in &stack {
-                    if let Some(m) = self.module_registry.get(m_path) {
-                        let mut mb = m.borrow_mut();
-                        mb.evaluated = true;
-                        mb.is_evaluating = false;
-                        if mb.error.is_none() {
-                            mb.error = Some(e.clone());
-                        }
-                    }
-                }
-                Err(e.clone())
-            }
-        }
-    }
-
-    pub(crate) fn get_object_property(
-        &mut self,
-        obj_id: u64,
-        key: &str,
-        this_val: &JsValue,
-    ) -> Completion {
-        // Check if object is a proxy
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            let key_val = self.symbol_key_to_jsvalue(key);
-            let receiver = this_val.clone();
-            match self.invoke_proxy_trap(obj_id, "get", vec![target_val.clone(), key_val, receiver])
-            {
-                Ok(Some(v)) => {
-                    // Invariant checks
-                    if let JsValue::Object(ref t) = target_val
-                        && let Some(tobj) = self.get_object(t.id)
-                    {
-                        let target_desc = tobj.borrow().get_own_property(key);
-                        if let Some(ref desc) = target_desc
-                            && desc.configurable == Some(false)
-                        {
-                            if desc.is_data_descriptor()
-                                && desc.writable == Some(false)
-                                && !same_value(
-                                    &v,
-                                    desc.value.as_ref().unwrap_or(&JsValue::Undefined),
-                                )
-                            {
-                                return Completion::Throw(self.create_type_error(
-                                        "'get' on proxy: property is a read-only and non-configurable data property on the proxy target but the proxy did not return its actual value",
-                                    ));
-                            }
-                            if desc.is_accessor_descriptor()
-                                && matches!(
-                                    desc.get.as_ref().unwrap_or(&JsValue::Undefined),
-                                    JsValue::Undefined
-                                )
-                                && !matches!(v, JsValue::Undefined)
-                            {
-                                return Completion::Throw(self.create_type_error(
-                                        "'get' on proxy: property is a non-configurable accessor property on the proxy target and does not have a getter function, but the trap did not return 'undefined'",
-                                    ));
-                            }
-                        }
-                    }
-                    return Completion::Normal(v);
-                }
-                Ok(None) => {
-                    // No trap, fall through to target
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.get_object_property(t.id, key, this_val);
-                    }
-                    return Completion::Normal(JsValue::Undefined);
-                }
-                Err(e) => return Completion::Throw(e),
-            }
-        }
-
-        // Module namespace: look up live binding from environment
-        {
-            let ns_data = self
-                .get_object(obj_id)
-                .and_then(|obj| obj.borrow().module_namespace.clone());
-            if let Some(ns_data) = ns_data {
-                // Deferred namespace: IsSymbolLikeNamespaceKey check
-                if ns_data.deferred
-                    && !Self::is_symbol_like_namespace_key(key, true)
-                    && let Err(e) = self.ensure_deferred_namespace_evaluation(obj_id)
-                {
-                    return Completion::Throw(e);
-                }
-                if let Some(binding_name) = ns_data.export_to_binding.get(key) {
-                    let module_path = ns_data.module_path.clone();
-                    match self.resolve_module_export_value(
-                        binding_name,
-                        &ns_data.env,
-                        module_path.as_deref(),
-                        key,
-                    ) {
-                        Ok(val) => return Completion::Normal(val),
-                        Err(e) => return Completion::Throw(e),
-                    }
-                }
-                // Fallback: check module's exports directly
-                if let Some(ref module_path) = ns_data.module_path
-                    && let Some(module) = self.module_registry.get(module_path)
-                    && let Some(val) = module.borrow().exports.get(key)
-                {
-                    return Completion::Normal(val.clone());
-                }
-            }
-        }
-
-        // TypedArray [[Get]]: canonical numeric index strings MUST NOT walk prototype
-        let is_typed_array_numeric = if let Some(obj) = self.get_object(obj_id) {
-            let b = obj.borrow();
-            b.typed_array_info.is_some()
-                && crate::interpreter::types::canonical_numeric_index_string(key).is_some()
-        } else {
-            false
-        };
-
-        // Check own property first, then walk prototype chain proxy-aware
-        let own_desc = if let Some(obj) = self.get_object(obj_id) {
-            obj.borrow().get_own_property_full(key)
-        } else {
-            None
-        };
-        match own_desc {
-            Some(ref d) if d.get.is_some() && !matches!(d.get, Some(JsValue::Undefined)) => {
-                let getter = d.get.clone().unwrap();
-                self.call_function(&getter, this_val, &[])
-            }
-            Some(ref d) if d.get.is_some() => Completion::Normal(JsValue::Undefined),
-            Some(ref d) => Completion::Normal(d.value.clone().unwrap_or(JsValue::Undefined)),
-            None => {
-                // TypedArray: numeric index strings must not walk prototype chain
-                if is_typed_array_numeric {
-                    return Completion::Normal(JsValue::Undefined);
-                }
-                // Walk prototype chain with proxy awareness
-                let proto = if let Some(obj) = self.get_object(obj_id) {
-                    obj.borrow().prototype.clone()
-                } else {
-                    None
-                };
-                if let Some(proto_rc) = proto {
-                    let proto_id = proto_rc.borrow().id.unwrap();
-                    self.get_object_property(proto_id, key, this_val)
-                } else {
-                    Completion::Normal(JsValue::Undefined)
-                }
-            }
-        }
-    }
-
-    /// Proxy-aware [[HasProperty]] - checks proxy `has` trap, recurses on target if no trap.
-    pub(crate) fn proxy_has_property(&mut self, obj_id: u64, key: &str) -> Result<bool, JsValue> {
-        if self.get_proxy_info(obj_id).is_some() {
-            let target_val = self.get_proxy_target_val(obj_id);
-            let key_val = self.symbol_key_to_jsvalue(key);
-            match self.invoke_proxy_trap(obj_id, "has", vec![target_val.clone(), key_val]) {
-                Ok(Some(v)) => {
-                    let trap_result = self.to_boolean_val(&v);
-                    if !trap_result
-                        && let JsValue::Object(ref t) = target_val
-                        && let Some(tobj) = self.get_object(t.id)
-                    {
-                        let target_desc = tobj.borrow().get_own_property(key);
-                        if let Some(ref desc) = target_desc {
-                            if desc.configurable == Some(false) {
-                                return Err(self.create_type_error(
-                                        "'has' on proxy: trap returned falsish for property which exists in the proxy target as non-configurable",
-                                    ));
-                            }
-                            if !tobj.borrow().extensible {
-                                return Err(self.create_type_error(
-                                        "'has' on proxy: trap returned falsish for property but the proxy target is not extensible",
-                                    ));
-                            }
-                        }
-                    }
-                    Ok(trap_result)
-                }
-                Ok(None) => {
-                    if let JsValue::Object(ref t) = target_val {
-                        return self.proxy_has_property(t.id, key);
-                    }
-                    Ok(false)
-                }
-                Err(e) => Err(e),
-            }
-        } else if let Some(obj) = self.get_object(obj_id) {
-            // Deferred namespace: trigger evaluation on [[HasProperty]] with non-symbol-like key
-            {
-                let is_deferred_ns = obj
-                    .borrow()
-                    .module_namespace
-                    .as_ref()
-                    .is_some_and(|ns| ns.deferred);
-                if is_deferred_ns && !Self::is_symbol_like_namespace_key(key, true) {
-                    self.ensure_deferred_namespace_evaluation(obj_id)?;
-                }
-            }
-            // TypedArray §10.4.5.3 [[HasProperty]]: numeric indices handled by IsValidIntegerIndex only
-            {
-                let b = obj.borrow();
-                if b.typed_array_info.is_some()
-                    && let Some(index) =
-                        crate::interpreter::types::canonical_numeric_index_string(key)
-                {
-                    return Ok(is_valid_integer_index(
-                        b.typed_array_info.as_ref().unwrap(),
-                        index,
-                    ));
-                }
-            }
-            if obj.borrow().has_own_property(key) {
-                return Ok(true);
-            }
-            // Walk prototype chain, checking for proxies
-            let proto = obj.borrow().prototype.clone();
-            if let Some(proto_rc) = proto {
-                let proto_id = proto_rc.borrow().id.unwrap();
-                return self.proxy_has_property(proto_id, key);
-            }
-            Ok(false)
-        } else {
-            Ok(false)
-        }
-    }
-
-    // === New with-scope reference semantics (spec-compliant) ===
-
-    /// Dynamically fetch @@unscopables from `obj_id` and check if `name` is blocked.
-    fn check_unscopables_dynamic(&mut self, obj_id: u64, name: &str) -> Result<bool, JsValue> {
-        let unscopables_val = {
-            let this_val = JsValue::Object(crate::types::JsObject { id: obj_id });
-            let key = "Symbol(Symbol.unscopables)";
-            match self.get_object_property(obj_id, key, &this_val) {
-                Completion::Normal(v) => v,
-                Completion::Throw(e) => return Err(e),
-                _ => JsValue::Undefined,
-            }
-        };
-        if let JsValue::Object(u_ref) = &unscopables_val {
-            let u_this = unscopables_val.clone();
-            match self.get_object_property(u_ref.id, name, &u_this) {
-                Completion::Normal(v) => Ok(self.to_boolean_val(&v)),
-                Completion::Throw(e) => Err(e),
-                _ => Ok(false),
-            }
-        } else {
-            Ok(false)
-        }
-    }
-
-    /// HasBinding for with-scopes: walks env chain, for each with-scope checks
-    /// proxy_has_property + check_unscopables_dynamic. Returns Ok(Some(obj_id)) if
-    /// the name resolves to a with-object, Ok(None) if found in a regular binding
-    /// or not found at all, Err on trap error.
-    pub(crate) fn resolve_with_has_binding(
-        &mut self,
-        name: &str,
-        env: &EnvRef,
-    ) -> Result<Option<u64>, JsValue> {
-        let mut current = Some(env.clone());
-        while let Some(env_ref) = current {
-            let env_borrow = env_ref.borrow();
-            if let Some(ref with) = env_borrow.with_object {
-                let obj_id = with.obj_id;
-                drop(env_borrow);
-                match self.proxy_has_property(obj_id, name) {
-                    Ok(true) => {
-                        if !self.check_unscopables_dynamic(obj_id, name)? {
-                            return Ok(Some(obj_id));
-                        }
-                    }
-                    Ok(false) => {}
-                    Err(e) => return Err(e),
-                }
-                let env_borrow = env_ref.borrow();
-                current = env_borrow.parent.clone();
-                continue;
-            }
-            if env_borrow.bindings.contains_key(name) {
-                return Ok(None);
-            }
-            if env_borrow.global_object.is_some() {
-                return Ok(None);
-            }
-            current = env_borrow.parent.clone();
-        }
-        Ok(None)
-    }
-
-    /// GetBindingValue for a known with-object: checks HasProperty(stillExists) + Get.
-    /// No unscopables check (already done in HasBinding).
-    fn with_get_binding_value(&mut self, obj_id: u64, name: &str, strict: bool) -> Completion {
-        match self.proxy_has_property(obj_id, name) {
-            Ok(true) => {
-                let this_val = JsValue::Object(crate::types::JsObject { id: obj_id });
-                self.get_object_property(obj_id, name, &this_val)
-            }
-            Ok(false) => {
-                if strict {
-                    Completion::Throw(
-                        self.create_reference_error(&format!("{name} is not defined")),
-                    )
-                } else {
-                    Completion::Normal(JsValue::Undefined)
-                }
-            }
-            Err(e) => Completion::Throw(e),
-        }
-    }
-
-    /// SetMutableBinding for a known with-object: checks HasProperty(stillExists) + Set.
-    /// No unscopables check (already done in HasBinding).
-    pub(crate) fn with_set_mutable_binding(
-        &mut self,
-        obj_id: u64,
-        name: &str,
-        value: JsValue,
-        strict: bool,
-    ) -> Result<(), JsValue> {
-        match self.proxy_has_property(obj_id, name) {
-            Ok(true) => {
-                let receiver = JsValue::Object(crate::types::JsObject { id: obj_id });
-                let success = self.proxy_set(obj_id, name, value, &receiver)?;
-                if !success && strict {
-                    return Err(self.create_type_error(&format!(
-                        "Cannot assign to read only property '{name}'"
-                    )));
-                }
-                Ok(())
-            }
-            Ok(false) => {
-                if strict {
-                    Err(self.create_reference_error(&format!("{name} is not defined")))
-                } else {
-                    let receiver = JsValue::Object(crate::types::JsObject { id: obj_id });
-                    self.proxy_set(obj_id, name, value, &receiver)?;
-                    Ok(())
-                }
-            }
-            Err(e) => Err(e),
-        }
-    }
-
     /// Resolve an identifier to a reference (for capturing before RHS evaluation).
     fn resolve_identifier_ref(
         &mut self,
@@ -16362,1261 +15151,6 @@ impl Interpreter {
         }
     }
 
-    fn eval_member(&mut self, obj: &Expression, prop: &MemberProperty, env: &EnvRef) -> Completion {
-        // §13.3.7.1: super[expr] — special evaluation order:
-        // 1. GetThisBinding (throws if uninitialized) — before key expression
-        // 2. Evaluate key expression
-        // 3. GetSuperBase (HomeObject.__proto__) — before ToPropertyKey
-        // 4. ToPropertyKey (in GetValue on the reference)
-        // 5. Property lookup on captured super base
-        if matches!(obj, Expression::Super) {
-            if let MemberProperty::Private(name) = prop {
-                if Self::this_is_in_tdz(env) {
-                    return Completion::Throw(self.create_reference_error(
-                        "Must call super constructor in derived class before accessing 'this' or returning from derived constructor",
-                    ));
-                }
-                let obj_val = match self.eval_expr(obj, env) {
-                    Completion::Normal(v) => v,
-                    other => return other,
-                };
-                let branded = self.resolve_private_name(name, env);
-                return match &obj_val {
-                    JsValue::Object(o) => {
-                        if let Some(obj_rc) = self.get_object(o.id) {
-                            let elem = obj_rc.borrow().private_fields.get(&branded).cloned();
-                            match elem {
-                                Some(PrivateElement::Field(v))
-                                | Some(PrivateElement::Method(v)) => Completion::Normal(v),
-                                Some(PrivateElement::Accessor { get, .. }) => {
-                                    if let Some(getter) = get {
-                                        self.call_function(&getter, &obj_val, &[])
-                                    } else {
-                                        Completion::Throw(self.create_type_error(&format!(
-                                            "Cannot read private member #{name} which has no getter"
-                                        )))
-                                    }
-                                }
-                                None => Completion::Throw(self.create_type_error(&format!(
-                                    "Cannot read private member #{name} from an object whose class did not declare it"
-                                ))),
-                            }
-                        } else {
-                            Completion::Normal(JsValue::Undefined)
-                        }
-                    }
-                    _ => Completion::Throw(self.create_type_error(&format!(
-                        "Cannot read private member #{name} from a non-object"
-                    ))),
-                };
-            }
-
-            // Step 2: GetThisBinding — throws ReferenceError if this is in TDZ
-            if Self::this_is_in_tdz(env) {
-                return Completion::Throw(self.create_reference_error(
-                    "Must call super constructor in derived class before accessing 'this' or returning from derived constructor",
-                ));
-            }
-            let this_val = env.borrow().get("this").unwrap_or(JsValue::Undefined);
-
-            // Steps 3-4: Evaluate key expression (without ToPropertyKey yet)
-            let raw_key = match prop {
-                MemberProperty::Dot(name) => {
-                    JsValue::String(crate::types::JsString::from_str(name))
-                }
-                MemberProperty::Computed(expr) => match self.eval_expr(expr, env) {
-                    Completion::Normal(v) => v,
-                    other => return other,
-                },
-                MemberProperty::Private(_) => unreachable!(),
-            };
-
-            // Step 7 → §13.3.7.3 step 3: GetSuperBase — capture BEFORE ToPropertyKey
-            let super_base_id = self.get_super_base_id(env);
-
-            // §6.2.5.5 GetValue step 3.c.i: ToPropertyKey (deferred until after GetSuperBase)
-            let key = match self.to_property_key(&raw_key) {
-                Ok(s) => s,
-                Err(e) => return Completion::Throw(e),
-            };
-
-            // Property lookup on captured super base
-            match super_base_id {
-                Some(base_id) => {
-                    return self.get_object_property(base_id, &key, &this_val);
-                }
-                None => {
-                    return Completion::Throw(self.create_type_error(&format!(
-                        "Cannot read properties of null (reading '{key}')"
-                    )));
-                }
-            }
-        }
-
-        let obj_val = match self.eval_expr(obj, env) {
-            Completion::Normal(v) => v,
-            other => return other,
-        };
-        if let MemberProperty::Private(name) = prop {
-            let branded = self.resolve_private_name(name, env);
-            return match &obj_val {
-                JsValue::Object(o) => {
-                    if let Some(obj) = self.get_object(o.id) {
-                        let elem = obj.borrow().private_fields.get(&branded).cloned();
-                        match elem {
-                            Some(PrivateElement::Field(v)) | Some(PrivateElement::Method(v)) => {
-                                Completion::Normal(v)
-                            }
-                            Some(PrivateElement::Accessor { get, .. }) => {
-                                if let Some(getter) = get {
-                                    self.call_function(&getter, &obj_val, &[])
-                                } else {
-                                    Completion::Throw(self.create_type_error(&format!(
-                                        "Cannot read private member #{name} which has no getter"
-                                    )))
-                                }
-                            }
-                            None => Completion::Throw(self.create_type_error(&format!(
-                                "Cannot read private member #{name} from an object whose class did not declare it"
-                            ))),
-                        }
-                    } else {
-                        Completion::Normal(JsValue::Undefined)
-                    }
-                }
-                _ => Completion::Throw(self.create_type_error(&format!(
-                    "Cannot read private member #{name} from a non-object"
-                ))),
-            };
-        }
-        // For computed properties, evaluate the expression but defer ToPropertyKey
-        // until after we check that the base is not null/undefined (spec: ToObject
-        // precedes ToPropertyKey per §6.2.5.5 GetValue step 3.a vs 3.c.i).
-        let (key, computed_raw) = match prop {
-            MemberProperty::Dot(name) => (name.clone(), None),
-            MemberProperty::Computed(expr) => {
-                let v = match self.eval_expr(expr, env) {
-                    Completion::Normal(v) => v,
-                    other => return other,
-                };
-                if matches!(&obj_val, JsValue::Null | JsValue::Undefined) {
-                    let err = self.create_type_error(&format!(
-                        "Cannot read properties of {obj_val} (reading property)"
-                    ));
-                    return Completion::Throw(err);
-                }
-                let key = match self.to_property_key(&v) {
-                    Ok(s) => s,
-                    Err(e) => return Completion::Throw(e),
-                };
-                (key, Some(()))
-            }
-            MemberProperty::Private(_) => unreachable!(),
-        };
-        let _ = computed_raw;
-        match &obj_val {
-            JsValue::Object(o) => self.get_object_property(o.id, &key, &obj_val.clone()),
-            JsValue::String(s) => {
-                if key == "length" {
-                    Completion::Normal(JsValue::Number(s.len() as f64))
-                } else if let Ok(idx) = key.parse::<usize>() {
-                    if idx < s.code_units.len() {
-                        Completion::Normal(JsValue::String(JsString {
-                            code_units: vec![s.code_units[idx]],
-                        }))
-                    } else {
-                        Completion::Normal(JsValue::Undefined)
-                    }
-                } else {
-                    let wrapper = match self.to_object(&obj_val) {
-                        Completion::Normal(v) => v,
-                        other => return other,
-                    };
-                    if let JsValue::Object(ref o) = wrapper {
-                        self.get_object_property(o.id, &key, &obj_val)
-                    } else {
-                        Completion::Normal(JsValue::Undefined)
-                    }
-                }
-            }
-            JsValue::Symbol(_) | JsValue::Number(_) | JsValue::Boolean(_) | JsValue::BigInt(_) => {
-                let wrapper = match self.to_object(&obj_val) {
-                    Completion::Normal(v) => v,
-                    other => return other,
-                };
-                if let JsValue::Object(ref o) = wrapper {
-                    self.get_object_property(o.id, &key, &obj_val)
-                } else {
-                    Completion::Normal(JsValue::Undefined)
-                }
-            }
-            JsValue::Undefined | JsValue::Null => {
-                let err = self.create_type_error(&format!(
-                    "Cannot read properties of {obj_val} (reading '{key}')"
-                ));
-                Completion::Throw(err)
-            }
-        }
-    }
-
-    fn eval_array_literal(&mut self, elements: &[Option<Expression>], env: &EnvRef) -> Completion {
-        let mut items: Vec<Option<JsValue>> = Vec::new();
-        for elem in elements {
-            match elem {
-                Some(Expression::Spread(inner)) => {
-                    let val = match self.eval_expr(inner, env) {
-                        Completion::Normal(v) => v,
-                        other => return other,
-                    };
-                    match self.iterate_to_vec(&val) {
-                        Ok(spread_items) => {
-                            for item in spread_items {
-                                items.push(Some(item));
-                            }
-                        }
-                        Err(e) => return Completion::Throw(e),
-                    }
-                }
-                Some(expr) => {
-                    let val = match self.eval_expr(expr, env) {
-                        Completion::Normal(v) => v,
-                        other => return other,
-                    };
-                    items.push(Some(val));
-                }
-                None => items.push(None), // elision — no own property
-            }
-        }
-        Completion::Normal(self.create_array_with_holes(items))
-    }
-
-    pub(crate) fn eval_class(
-        &mut self,
-        name: &str,
-        class_binding_name: &str,
-        super_class: &Option<Box<Expression>>,
-        body: &[ClassElement],
-        env: &EnvRef,
-        class_source_text: Option<String>,
-    ) -> Completion {
-        let brand_id = self.next_class_brand_id;
-        self.next_class_brand_id += 1;
-        let mut pn_set = std::collections::HashMap::new();
-        for elem in body {
-            match elem {
-                ClassElement::Method(m) => {
-                    if let PropertyKey::Private(n) = &m.key {
-                        pn_set
-                            .entry(n.clone())
-                            .or_insert_with(|| format!("{n}#{brand_id}"));
-                    }
-                }
-                ClassElement::Property(p) => {
-                    if let PropertyKey::Private(n) = &p.key {
-                        pn_set
-                            .entry(n.clone())
-                            .or_insert_with(|| format!("{n}#{brand_id}"));
-                    }
-                }
-                ClassElement::StaticBlock(_) => {}
-            }
-        }
-        self.class_private_names.push(pn_set);
-        let result = self.eval_class_inner(
-            name,
-            class_binding_name,
-            super_class,
-            body,
-            env,
-            class_source_text,
-        );
-        self.class_private_names.pop();
-        result
-    }
-
-    fn eval_class_inner(
-        &mut self,
-        name: &str,
-        class_binding_name: &str,
-        super_class: &Option<Box<Expression>>,
-        body: &[ClassElement],
-        env: &EnvRef,
-        class_source_text: Option<String>,
-    ) -> Completion {
-        // Find constructor method
-        let ctor_method = body.iter().find_map(|elem| {
-            if let ClassElement::Method(m) = elem
-                && m.kind == ClassMethodKind::Constructor
-            {
-                return Some(m);
-            }
-            None
-        });
-
-        // Per spec §15.7.14: Create class environment FIRST so heritage expression
-        // is evaluated in it, and closures in heritage capture the class name binding.
-        let class_env = Environment::new(Some(env.clone()));
-        class_env.borrow_mut().class_private_names = self.class_private_names.last().cloned();
-        class_env.borrow_mut().strict = true;
-        // Pre-declare class name as uninitialized immutable binding (spec step 4a)
-        if !class_binding_name.is_empty() {
-            class_env
-                .borrow_mut()
-                .declare(class_binding_name, BindingKind::Const);
-        }
-
-        // Evaluate super class in class_env context (spec step 6a-6b)
-        let super_val = if let Some(sc) = super_class {
-            match self.eval_expr(sc, &class_env) {
-                Completion::Normal(v) => Some(v),
-                other => return other,
-            }
-        } else {
-            None
-        };
-
-        // Validate super class: must be null or a constructor
-        if let Some(ref sv) = super_val
-            && !matches!(sv, JsValue::Null)
-            && !self.is_constructor(sv)
-        {
-            return Completion::Throw(
-                self.create_type_error("Class extends value is not a constructor or null"),
-            );
-        }
-
-        if let Some(ref sv) = super_val {
-            class_env
-                .borrow_mut()
-                .declare("__super__", BindingKind::Const);
-            class_env
-                .borrow_mut()
-                .initialize_binding("__super__", sv.clone());
-        }
-
-        // Create constructor function (classes are always strict mode)
-        let ctor_func = if let Some(cm) = ctor_method {
-            JsFunction::User {
-                name: Some(name.to_string()),
-                params: cm.value.params.clone(),
-                body: cm.value.body.clone(),
-                closure: class_env.clone(),
-                is_arrow: false,
-                is_strict: true,
-                is_generator: false,
-                is_async: false,
-                is_method: false,
-                source_text: class_source_text.clone(),
-                captured_new_target: None,
-            }
-        } else if super_val.is_some() {
-            JsFunction::User {
-                name: Some(name.to_string()),
-                params: vec![Pattern::Rest(Box::new(Pattern::Identifier("args".into())))],
-                body: vec![Statement::Expression(Expression::Call(
-                    Box::new(Expression::Super),
-                    vec![Expression::Spread(Box::new(Expression::Identifier(
-                        "args".into(),
-                    )))],
-                ))],
-                closure: class_env.clone(),
-                is_arrow: false,
-                is_strict: true,
-                is_generator: false,
-                is_async: false,
-                is_method: false,
-                source_text: class_source_text.clone(),
-                captured_new_target: None,
-            }
-        } else {
-            JsFunction::User {
-                name: Some(name.to_string()),
-                params: vec![],
-                body: vec![],
-                closure: class_env.clone(),
-                is_arrow: false,
-                is_strict: true,
-                is_generator: false,
-                is_async: false,
-                is_method: false,
-                source_text: class_source_text.clone(),
-                captured_new_target: None,
-            }
-        };
-
-        let ctor_val = self.create_function(ctor_func);
-
-        // Mark derived class constructors and make .prototype writable:false
-        if let JsValue::Object(ref o) = ctor_val
-            && let Some(func_obj) = self.get_object(o.id)
-        {
-            func_obj.borrow_mut().is_class_constructor = true;
-            if super_val.is_some() {
-                func_obj.borrow_mut().is_derived_class_constructor = true;
-                if ctor_method.is_none() {
-                    func_obj.borrow_mut().is_default_derived_constructor = true;
-                }
-            }
-            // Per spec §14.6.13: class .prototype is {writable: false, enumerable: false, configurable: false}
-            let proto_val_for_desc = func_obj.borrow().get_property("prototype");
-            func_obj.borrow_mut().insert_property(
-                "prototype".to_string(),
-                PropertyDescriptor::data(proto_val_for_desc, false, false, false),
-            );
-        }
-
-        // Initialize class name binding (pre-declared above; spec §15.7.14 step 18.c/26.d)
-        if !class_binding_name.is_empty() {
-            class_env
-                .borrow_mut()
-                .initialize_binding(class_binding_name, ctor_val.clone());
-        }
-
-        // Store constructor func for dynamic GetSuperConstructor (§13.3.7.2)
-        if super_val.is_some() {
-            class_env
-                .borrow_mut()
-                .declare("__constructor_func__", BindingKind::Const);
-            class_env
-                .borrow_mut()
-                .initialize_binding("__constructor_func__", ctor_val.clone());
-        }
-
-        // Get the prototype object that was auto-created by create_function
-        let proto_obj = if let JsValue::Object(ref o) = ctor_val {
-            if let Some(func_obj) = self.get_object(o.id) {
-                let proto_val = func_obj.borrow().get_property("prototype");
-                if let JsValue::Object(ref p) = proto_val {
-                    self.get_object(p.id)
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        } else {
-            None
-        };
-
-        // Set up inheritance
-        if let Some(ref sv) = super_val
-            && let JsValue::Object(super_o) = sv
-        {
-            let super_o_id = super_o.id;
-            let sv_clone = sv.clone();
-            // Step 5.e: proto.[[Prototype]] = superclass.prototype
-            // Must use Get() to invoke accessor properties (e.g. getter-defined prototype)
-            let super_proto_val = match self.get_object_property(super_o_id, "prototype", &sv_clone)
-            {
-                Completion::Normal(v) => v,
-                other => return other,
-            };
-            // Validate: must be Object or null
-            match &super_proto_val {
-                JsValue::Object(_) | JsValue::Null => {}
-                _ => {
-                    return Completion::Throw(self.create_type_error(
-                        "Class extends value does not have valid prototype property",
-                    ));
-                }
-            }
-            if let JsValue::Object(ref sp) = super_proto_val
-                && let Some(super_proto) = self.get_object(sp.id)
-                && let Some(ref proto) = proto_obj
-            {
-                proto.borrow_mut().prototype = Some(super_proto);
-            }
-            // Step 7.a: F.[[Prototype]] = superclass (for static method inheritance)
-            if let JsValue::Object(ref o) = ctor_val
-                && let Some(ctor_obj) = self.get_object(o.id)
-                && let Some(super_obj) = self.get_object(super_o_id)
-            {
-                ctor_obj.borrow_mut().prototype = Some(super_obj);
-            }
-        }
-
-        // Handle `extends null` — set prototype's [[Prototype]] to null
-        if let Some(JsValue::Null) = super_val
-            && let Some(ref proto) = proto_obj
-        {
-            proto.borrow_mut().prototype = None;
-        }
-
-        // Set __home_object__ in class_env for the constructor (which uses class_env as
-        // its closure directly). Non-constructor methods get per-method closures that
-        // shadow this with their own __home_object__ binding.
-        let ctor_home = if let Some(ref p) = proto_obj {
-            let pid = p.borrow().id.unwrap();
-            JsValue::Object(crate::types::JsObject { id: pid })
-        } else {
-            JsValue::Undefined
-        };
-        class_env
-            .borrow_mut()
-            .declare("__home_object__", BindingKind::Const);
-        class_env
-            .borrow_mut()
-            .initialize_binding("__home_object__", ctor_home);
-
-        // Create environment for static field initializers with `this` = constructor
-        let static_field_env = Environment::new_function_scope(Some(class_env.clone()));
-        {
-            let mut sfe = static_field_env.borrow_mut();
-            sfe.bindings.insert(
-                "this".to_string(),
-                Binding {
-                    value: ctor_val.clone(),
-                    kind: BindingKind::Const,
-                    initialized: true,
-                    deletable: false,
-                },
-            );
-            sfe.is_field_initializer = true;
-            sfe.class_private_names = self.class_private_names.last().cloned();
-            // Set __home_object__ for super property access in static field initializers.
-            // Static field HomeObject = constructor.
-            sfe.bindings.insert(
-                "__home_object__".to_string(),
-                Binding {
-                    value: ctor_val.clone(),
-                    kind: BindingKind::Const,
-                    initialized: true,
-                    deletable: false,
-                },
-            );
-        }
-
-        // Per spec §15.7.14 step 28-34: Process all elements in two phases.
-        // Phase 1: Evaluate ALL computed keys in declaration order, install methods,
-        //          collect instance field defs, and defer static fields/blocks.
-        // Phase 2: Execute static field initializers and static blocks in order.
-        enum DeferredStatic {
-            PublicField(String, Option<Expression>),
-            // (source_name, branded_name, initializer)
-            PrivateField(String, String, Option<Expression>),
-            Block(Vec<Statement>),
-        }
-        let mut deferred_static: Vec<DeferredStatic> = Vec::new();
-
-        for elem in body {
-            match elem {
-                ClassElement::Method(m) => {
-                    if m.kind == ClassMethodKind::Constructor {
-                        continue;
-                    }
-                    let (key, fn_name_for_key) = match &m.key {
-                        PropertyKey::Identifier(s) | PropertyKey::String(s) => {
-                            (s.clone(), s.clone())
-                        }
-                        PropertyKey::Number(n) => {
-                            let s = to_js_string(&JsValue::Number(*n));
-                            (s.clone(), s)
-                        }
-                        PropertyKey::Computed(expr) => match self.eval_expr(expr, &class_env) {
-                            Completion::Normal(v) => {
-                                let is_symbol = matches!(&v, JsValue::Symbol(_));
-                                let fn_name = if let JsValue::Symbol(ref sym) = v {
-                                    match &sym.description {
-                                        Some(desc) => format!("[{}]", desc),
-                                        None => String::new(),
-                                    }
-                                } else {
-                                    String::new()
-                                };
-                                match self.to_property_key(&v) {
-                                    Ok(s) => {
-                                        let name = if is_symbol { fn_name } else { s.clone() };
-                                        (s, name)
-                                    }
-                                    Err(e) => return Completion::Throw(e),
-                                }
-                            }
-                            other => return other,
-                        },
-                        PropertyKey::Private(name) => {
-                            let branded = self.resolve_private_name(name, &class_env);
-                            let priv_home_target = if m.is_static {
-                                ctor_val.clone()
-                            } else if let Some(ref p) = proto_obj {
-                                let pid = p.borrow().id.unwrap();
-                                JsValue::Object(crate::types::JsObject { id: pid })
-                            } else {
-                                JsValue::Undefined
-                            };
-                            let method_closure = Environment::new(Some(class_env.clone()));
-                            method_closure
-                                .borrow_mut()
-                                .declare("__home_object__", BindingKind::Const);
-                            method_closure
-                                .borrow_mut()
-                                .initialize_binding("__home_object__", priv_home_target);
-                            let method_func = JsFunction::User {
-                                name: Some(format!("#{name}")),
-                                params: m.value.params.clone(),
-                                body: m.value.body.clone(),
-                                closure: method_closure,
-                                is_arrow: false,
-                                is_strict: true,
-                                is_generator: m.value.is_generator,
-                                is_async: m.value.is_async,
-                                is_method: true,
-                                source_text: m.value.source_text.clone(),
-                                captured_new_target: None,
-                            };
-                            let method_val = self.create_function(method_func);
-
-                            if m.is_static {
-                                if let JsValue::Object(ref o) = ctor_val
-                                    && let Some(func_obj) = self.get_object(o.id)
-                                {
-                                    match m.kind {
-                                        ClassMethodKind::Get => {
-                                            let existing = func_obj
-                                                .borrow()
-                                                .private_fields
-                                                .get(&branded)
-                                                .cloned();
-                                            let elem = if let Some(PrivateElement::Accessor {
-                                                get: _,
-                                                set,
-                                            }) = existing
-                                            {
-                                                PrivateElement::Accessor {
-                                                    get: Some(method_val),
-                                                    set,
-                                                }
-                                            } else {
-                                                PrivateElement::Accessor {
-                                                    get: Some(method_val),
-                                                    set: None,
-                                                }
-                                            };
-                                            func_obj
-                                                .borrow_mut()
-                                                .private_fields
-                                                .insert(branded.clone(), elem);
-                                        }
-                                        ClassMethodKind::Set => {
-                                            let existing = func_obj
-                                                .borrow()
-                                                .private_fields
-                                                .get(&branded)
-                                                .cloned();
-                                            let elem = if let Some(PrivateElement::Accessor {
-                                                get,
-                                                set: _,
-                                            }) = existing
-                                            {
-                                                PrivateElement::Accessor {
-                                                    get,
-                                                    set: Some(method_val),
-                                                }
-                                            } else {
-                                                PrivateElement::Accessor {
-                                                    get: None,
-                                                    set: Some(method_val),
-                                                }
-                                            };
-                                            func_obj
-                                                .borrow_mut()
-                                                .private_fields
-                                                .insert(branded.clone(), elem);
-                                        }
-                                        _ => {
-                                            func_obj.borrow_mut().private_fields.insert(
-                                                branded.clone(),
-                                                PrivateElement::Method(method_val),
-                                            );
-                                        }
-                                    }
-                                }
-                            } else if let JsValue::Object(ref o) = ctor_val
-                                && let Some(func_obj) = self.get_object(o.id)
-                            {
-                                match m.kind {
-                                    ClassMethodKind::Get => {
-                                        let mut b = func_obj.borrow_mut();
-                                        let mut found = false;
-                                        for idef in b.class_instance_field_defs.iter_mut() {
-                                            if let InstanceFieldDef::Private(
-                                                PrivateFieldDef::Accessor {
-                                                    name: n, get: g, ..
-                                                },
-                                            ) = idef
-                                                && *n == branded
-                                            {
-                                                *g = Some(method_val.clone());
-                                                found = true;
-                                                break;
-                                            }
-                                        }
-                                        if !found {
-                                            b.class_instance_field_defs.push(
-                                                InstanceFieldDef::Private(
-                                                    PrivateFieldDef::Accessor {
-                                                        name: branded.clone(),
-                                                        get: Some(method_val),
-                                                        set: None,
-                                                    },
-                                                ),
-                                            );
-                                        }
-                                    }
-                                    ClassMethodKind::Set => {
-                                        let mut b = func_obj.borrow_mut();
-                                        let mut found = false;
-                                        for idef in b.class_instance_field_defs.iter_mut() {
-                                            if let InstanceFieldDef::Private(
-                                                PrivateFieldDef::Accessor {
-                                                    name: n, set: s, ..
-                                                },
-                                            ) = idef
-                                                && *n == branded
-                                            {
-                                                *s = Some(method_val.clone());
-                                                found = true;
-                                                break;
-                                            }
-                                        }
-                                        if !found {
-                                            b.class_instance_field_defs.push(
-                                                InstanceFieldDef::Private(
-                                                    PrivateFieldDef::Accessor {
-                                                        name: branded.clone(),
-                                                        get: None,
-                                                        set: Some(method_val),
-                                                    },
-                                                ),
-                                            );
-                                        }
-                                    }
-                                    _ => {
-                                        func_obj.borrow_mut().class_instance_field_defs.push(
-                                            InstanceFieldDef::Private(PrivateFieldDef::Method {
-                                                name: branded.clone(),
-                                                value: method_val,
-                                            }),
-                                        );
-                                    }
-                                }
-                            }
-                            continue;
-                        }
-                    };
-                    let method_display_name = match m.kind {
-                        ClassMethodKind::Get => format!("get {fn_name_for_key}"),
-                        ClassMethodKind::Set => format!("set {fn_name_for_key}"),
-                        _ => fn_name_for_key.clone(),
-                    };
-                    let home_target = if m.is_static {
-                        ctor_val.clone()
-                    } else if let Some(ref p) = proto_obj {
-                        let pid = p.borrow().id.unwrap();
-                        JsValue::Object(crate::types::JsObject { id: pid })
-                    } else {
-                        JsValue::Undefined
-                    };
-                    let method_closure = Environment::new(Some(class_env.clone()));
-                    method_closure
-                        .borrow_mut()
-                        .declare("__home_object__", BindingKind::Const);
-                    method_closure
-                        .borrow_mut()
-                        .initialize_binding("__home_object__", home_target);
-                    let method_func = JsFunction::User {
-                        name: Some(method_display_name),
-                        params: m.value.params.clone(),
-                        body: m.value.body.clone(),
-                        closure: method_closure,
-                        is_arrow: false,
-                        is_strict: true,
-                        is_generator: m.value.is_generator,
-                        is_async: m.value.is_async,
-                        is_method: true,
-                        source_text: m.value.source_text.clone(),
-                        captured_new_target: None,
-                    };
-                    let method_val = self.create_function(method_func);
-
-                    let target = if m.is_static {
-                        if let JsValue::Object(ref o) = ctor_val {
-                            self.get_object(o.id)
-                        } else {
-                            None
-                        }
-                    } else {
-                        proto_obj.clone()
-                    };
-                    if let Some(ref t) = target {
-                        let ok = match m.kind {
-                            ClassMethodKind::Get => {
-                                let mut desc = t.borrow().properties.get(&key).cloned().unwrap_or(
-                                    PropertyDescriptor {
-                                        value: None,
-                                        writable: None,
-                                        get: None,
-                                        set: None,
-                                        enumerable: Some(false),
-                                        configurable: Some(true),
-                                    },
-                                );
-                                desc.get = Some(method_val);
-                                desc.value = None;
-                                desc.writable = None;
-                                t.borrow_mut().define_own_property(key, desc)
-                            }
-                            ClassMethodKind::Set => {
-                                let mut desc = t.borrow().properties.get(&key).cloned().unwrap_or(
-                                    PropertyDescriptor {
-                                        value: None,
-                                        writable: None,
-                                        get: None,
-                                        set: None,
-                                        enumerable: Some(false),
-                                        configurable: Some(true),
-                                    },
-                                );
-                                desc.set = Some(method_val);
-                                desc.value = None;
-                                desc.writable = None;
-                                t.borrow_mut().define_own_property(key, desc)
-                            }
-                            _ => {
-                                let desc = PropertyDescriptor::data(method_val, true, false, true);
-                                t.borrow_mut().define_own_property(key, desc)
-                            }
-                        };
-                        if !ok {
-                            return Completion::Throw(
-                                self.create_type_error("Cannot redefine non-configurable property"),
-                            );
-                        }
-                    }
-                }
-                ClassElement::Property(p) => {
-                    // Check if this is a private field
-                    if let PropertyKey::Private(name) = &p.key {
-                        let branded = self.resolve_private_name(name, &class_env);
-                        if !p.is_static {
-                            // Store instance private field definition
-                            if let JsValue::Object(ref o) = ctor_val
-                                && let Some(func_obj) = self.get_object(o.id)
-                            {
-                                func_obj.borrow_mut().class_instance_field_defs.push(
-                                    InstanceFieldDef::Private(PrivateFieldDef::Field {
-                                        name: branded,
-                                        initializer: p.value.clone(),
-                                    }),
-                                );
-                            }
-                        } else {
-                            // Defer static private field initializer to phase 2
-                            deferred_static.push(DeferredStatic::PrivateField(
-                                name.clone(),
-                                branded,
-                                p.value.clone(),
-                            ));
-                        }
-                        continue;
-                    }
-                    if p.is_static {
-                        // Evaluate computed key NOW in phase 1, defer initializer to phase 2
-                        let key = match &p.key {
-                            PropertyKey::Identifier(s) | PropertyKey::String(s) => s.clone(),
-                            PropertyKey::Number(n) => to_js_string(&JsValue::Number(*n)),
-                            PropertyKey::Computed(expr) => match self.eval_expr(expr, &class_env) {
-                                Completion::Normal(v) => match self.to_property_key(&v) {
-                                    Ok(s) => s,
-                                    Err(e) => return Completion::Throw(e),
-                                },
-                                other => return other,
-                            },
-                            PropertyKey::Private(_) => unreachable!(),
-                        };
-                        if key == "prototype" {
-                            return Completion::Throw(self.create_type_error(
-                                "Classes may not have a static property named 'prototype'",
-                            ));
-                        }
-                        deferred_static.push(DeferredStatic::PublicField(key, p.value.clone()));
-                    } else {
-                        // Instance field: evaluate computed key, store field def
-                        let key = match &p.key {
-                            PropertyKey::Identifier(s) | PropertyKey::String(s) => s.clone(),
-                            PropertyKey::Number(n) => to_js_string(&JsValue::Number(*n)),
-                            PropertyKey::Computed(expr) => match self.eval_expr(expr, &class_env) {
-                                Completion::Normal(v) => match self.to_property_key(&v) {
-                                    Ok(s) => s,
-                                    Err(e) => return Completion::Throw(e),
-                                },
-                                other => return other,
-                            },
-                            PropertyKey::Private(_) => unreachable!(),
-                        };
-                        if let JsValue::Object(ref o) = ctor_val
-                            && let Some(func_obj) = self.get_object(o.id)
-                        {
-                            func_obj
-                                .borrow_mut()
-                                .class_instance_field_defs
-                                .push(InstanceFieldDef::Public(key, p.value.clone()));
-                        }
-                    }
-                }
-                ClassElement::StaticBlock(stmts) => {
-                    // Defer static block execution to phase 2
-                    deferred_static.push(DeferredStatic::Block(stmts.clone()));
-                }
-            }
-        }
-
-        // Phase 2: Execute deferred static field initializers and static blocks
-        for deferred in deferred_static {
-            match deferred {
-                DeferredStatic::PublicField(key, initializer) => {
-                    let val = if let Some(ref expr) = initializer {
-                        match self.eval_expr(expr, &static_field_env) {
-                            Completion::Normal(v) => {
-                                if expr.is_anonymous_function_definition() {
-                                    self.set_function_name(&v, &key);
-                                }
-                                v
-                            }
-                            other => return other,
-                        }
-                    } else {
-                        JsValue::Undefined
-                    };
-                    if let JsValue::Object(ref o) = ctor_val
-                        && let Some(func_obj) = self.get_object(o.id)
-                    {
-                        func_obj.borrow_mut().insert_value(key, val);
-                    }
-                }
-                DeferredStatic::PrivateField(source_name, branded, initializer) => {
-                    let display_name = format!("#{source_name}");
-                    let val = if let Some(ref expr) = initializer {
-                        match self.eval_expr(expr, &static_field_env) {
-                            Completion::Normal(v) => {
-                                if expr.is_anonymous_function_definition() {
-                                    self.set_function_name(&v, &display_name);
-                                }
-                                v
-                            }
-                            other => return other,
-                        }
-                    } else {
-                        JsValue::Undefined
-                    };
-                    if let JsValue::Object(ref o) = ctor_val
-                        && let Some(func_obj) = self.get_object(o.id)
-                    {
-                        if !func_obj.borrow().extensible {
-                            return Completion::Throw(self.create_type_error(
-                                "Cannot add private field to non-extensible object",
-                            ));
-                        }
-                        if func_obj.borrow().private_fields.contains_key(&branded) {
-                            return Completion::Throw(self.create_type_error(
-                                "Cannot initialize private field twice on the same object",
-                            ));
-                        }
-                        func_obj
-                            .borrow_mut()
-                            .private_fields
-                            .insert(branded, PrivateElement::Field(val));
-                    }
-                }
-                DeferredStatic::Block(stmts) => {
-                    let block_env = Environment::new_function_scope(Some(class_env.clone()));
-                    {
-                        let mut be = block_env.borrow_mut();
-                        be.strict = true; // class bodies are always strict
-                        be.bindings.insert(
-                            "this".to_string(),
-                            Binding {
-                                value: ctor_val.clone(),
-                                kind: BindingKind::Const,
-                                initialized: true,
-                                deletable: false,
-                            },
-                        );
-                        be.bindings.insert(
-                            "__home_object__".to_string(),
-                            Binding {
-                                value: ctor_val.clone(),
-                                kind: BindingKind::Const,
-                                initialized: true,
-                                deletable: false,
-                            },
-                        );
-                        be.class_private_names = self.class_private_names.last().cloned();
-                    }
-                    let result = self.exec_statements(&stmts, &block_env);
-                    let result = self.dispose_resources(&block_env, result);
-                    match result {
-                        Completion::Normal(_) => {}
-                        Completion::Throw(e) => return Completion::Throw(e),
-                        _ => {}
-                    }
-                }
-            }
-        }
-
-        Completion::Normal(ctor_val)
-    }
-
-    /// CopyDataProperties (§7.3.26) — copies own enumerable properties from source
-    /// to target obj_data. Properly handles Proxy traps and Symbol keys.
-    pub(crate) fn copy_data_properties(
-        &mut self,
-        src_id: u64,
-        src_val: &JsValue,
-        excluded: &[String],
-    ) -> Result<Vec<(String, JsValue)>, JsValue> {
-        let mut result = Vec::new();
-        let keys = self.proxy_own_keys(src_id)?;
-        for key_val in keys {
-            let key_str = self.to_property_key(&key_val)?;
-            if excluded.contains(&key_str) {
-                continue;
-            }
-            let is_enumerable = if self.get_proxy_info(src_id).is_some() {
-                let target_proxy_val = self.get_proxy_target_val(src_id);
-                match self.invoke_proxy_trap(
-                    src_id,
-                    "getOwnPropertyDescriptor",
-                    vec![target_proxy_val, key_val.clone()],
-                ) {
-                    Ok(Some(v)) => {
-                        if v.is_undefined() {
-                            continue;
-                        }
-                        if let JsValue::Object(ref dobj) = v
-                            && let Some(desc_rc) = self.get_object(dobj.id)
-                        {
-                            match desc_rc.borrow().get_property_value("enumerable") {
-                                Some(ev) => self.to_boolean_val(&ev),
-                                None => false,
-                            }
-                        } else {
-                            continue;
-                        }
-                    }
-                    Ok(None) => {
-                        if let Some(obj) = self.get_object(src_id) {
-                            let desc = obj.borrow().get_own_property(&key_str);
-                            match desc {
-                                Some(d) => d.enumerable != Some(false),
-                                None => continue,
-                            }
-                        } else {
-                            continue;
-                        }
-                    }
-                    Err(e) => return Err(e),
-                }
-            } else if let Some(obj) = self.get_object(src_id) {
-                let desc = obj.borrow().get_own_property(&key_str);
-                match desc {
-                    Some(d) => d.enumerable != Some(false),
-                    None => continue,
-                }
-            } else {
-                continue;
-            };
-            if !is_enumerable {
-                continue;
-            }
-            let val = match self.get_object_property(src_id, &key_str, src_val) {
-                Completion::Normal(v) => v,
-                Completion::Throw(e) => return Err(e),
-                _ => JsValue::Undefined,
-            };
-            result.push((key_str, val));
-        }
-        Ok(result)
-    }
-
-    fn eval_object_literal(&mut self, props: &[Property], env: &EnvRef) -> Completion {
-        let mut obj_data = JsObjectData::new();
-        obj_data.prototype = self.realm().object_prototype.clone();
-        let mut method_values: Vec<JsValue> = Vec::new();
-        for prop in props {
-            let (key, fn_name_for_key) = match &prop.key {
-                PropertyKey::Identifier(n) => (n.clone(), n.clone()),
-                PropertyKey::String(s) => (s.clone(), s.clone()),
-                PropertyKey::Number(n) => {
-                    let s = number_ops::to_string(*n);
-                    (s.clone(), s)
-                }
-                PropertyKey::Computed(expr) => {
-                    let v = match self.eval_expr(expr, env) {
-                        Completion::Normal(v) => v,
-                        other => return other,
-                    };
-                    let is_symbol = matches!(&v, JsValue::Symbol(_));
-                    let fn_name = if let JsValue::Symbol(ref sym) = v {
-                        match &sym.description {
-                            Some(desc) => format!("[{}]", desc),
-                            None => String::new(),
-                        }
-                    } else {
-                        String::new()
-                    };
-                    let pk = match self.to_property_key(&v) {
-                        Ok(s) => s,
-                        Err(e) => return Completion::Throw(e),
-                    };
-                    let name = if is_symbol { fn_name } else { pk.clone() };
-                    (pk, name)
-                }
-                PropertyKey::Private(_) => {
-                    return Completion::Throw(
-                        self.create_type_error("Private names are not valid in object literals"),
-                    );
-                }
-            };
-            if prop.method {
-                self.next_function_is_method = true;
-            }
-            let value = match self.eval_expr(&prop.value, env) {
-                Completion::Normal(v) => v,
-                other => {
-                    self.next_function_is_method = false;
-                    return other;
-                }
-            };
-            self.next_function_is_method = false;
-            // Handle spread — CopyDataProperties (§7.3.26)
-            if let Expression::Spread(inner) = &prop.value {
-                let spread_val = match self.eval_expr(inner, env) {
-                    Completion::Normal(v) => v,
-                    other => return other,
-                };
-                if let JsValue::Object(ref o) = spread_val {
-                    let src_id = o.id;
-                    match self.copy_data_properties(src_id, &spread_val, &[]) {
-                        Ok(pairs) => {
-                            for (k, v) in pairs {
-                                obj_data.insert_value(k, v);
-                            }
-                        }
-                        Err(e) => return Completion::Throw(e),
-                    }
-                }
-                continue;
-            }
-            match prop.kind {
-                PropertyKind::Get => {
-                    self.set_function_name(&value, &format!("get {fn_name_for_key}"));
-                    method_values.push(value.clone());
-                    let mut desc =
-                        obj_data
-                            .properties
-                            .get(&key)
-                            .cloned()
-                            .unwrap_or(PropertyDescriptor {
-                                value: None,
-                                writable: None,
-                                get: None,
-                                set: None,
-                                enumerable: Some(true),
-                                configurable: Some(true),
-                            });
-                    desc.get = Some(value);
-                    desc.value = None;
-                    desc.writable = None;
-                    obj_data.insert_property(key, desc);
-                }
-                PropertyKind::Set => {
-                    self.set_function_name(&value, &format!("set {fn_name_for_key}"));
-                    method_values.push(value.clone());
-                    let mut desc =
-                        obj_data
-                            .properties
-                            .get(&key)
-                            .cloned()
-                            .unwrap_or(PropertyDescriptor {
-                                value: None,
-                                writable: None,
-                                get: None,
-                                set: None,
-                                enumerable: Some(true),
-                                configurable: Some(true),
-                            });
-                    desc.set = Some(value);
-                    desc.value = None;
-                    desc.writable = None;
-                    obj_data.insert_property(key, desc);
-                }
-                _ => {
-                    // __proto__: value sets [[Prototype]] per spec §13.2.5.5
-                    // Only plain property init, not methods, computed, or shorthand
-                    if key == "__proto__" && !prop.computed && !prop.shorthand && !prop.method {
-                        match &value {
-                            JsValue::Object(o) => {
-                                obj_data.prototype = self.get_object(o.id);
-                            }
-                            JsValue::Null => {
-                                obj_data.prototype = None;
-                            }
-                            _ => {
-                                // Non-object, non-null values are ignored per spec
-                            }
-                        }
-                    } else {
-                        if prop.value.is_anonymous_function_definition() {
-                            self.set_function_name(&value, &fn_name_for_key);
-                        }
-                        if prop.method {
-                            method_values.push(value.clone());
-                        }
-                        obj_data.insert_value(key, value);
-                    }
-                }
-            }
-        }
-        let obj = Rc::new(RefCell::new(obj_data));
-        let id = self.allocate_object_slot(obj);
-        // Set __home_object__ for concise methods, getters, and setters
-        let obj_val = JsValue::Object(crate::types::JsObject { id });
-        {
-            for val in &method_values {
-                if let JsValue::Object(fo) = val
-                    && let Some(func_obj) = self.get_object(fo.id)
-                {
-                    let old_closure = if let Some(JsFunction::User { ref closure, .. }) =
-                        func_obj.borrow().callable
-                    {
-                        Some(closure.clone())
-                    } else {
-                        None
-                    };
-                    if let Some(old_closure) = old_closure {
-                        let wrapper = Environment::new(Some(old_closure));
-                        wrapper
-                            .borrow_mut()
-                            .declare("__home_object__", BindingKind::Const);
-                        wrapper
-                            .borrow_mut()
-                            .initialize_binding("__home_object__", obj_val.clone());
-                        if let Some(JsFunction::User {
-                            ref mut closure, ..
-                        }) = func_obj.borrow_mut().callable
-                        {
-                            *closure = wrapper;
-                        }
-                    }
-                    // Methods must not have own caller/arguments (spec §15.4)
-                    func_obj.borrow_mut().properties.remove("caller");
-                    func_obj.borrow_mut().properties.remove("arguments");
-                }
-            }
-        }
-        Completion::Normal(obj_val)
-    }
-
     fn call_async_function(
         &mut self,
         params: &[Pattern],
@@ -18939,419 +16473,6 @@ impl Interpreter {
             Some(Ok(v)) => Completion::Normal(v),
             Some(Err(e)) => Completion::Throw(e),
             None => Completion::Normal(JsValue::Undefined),
-        }
-    }
-
-    fn dynamic_import(&mut self, specifier: &str) -> Completion {
-        let resolved =
-            match self.resolve_module_specifier(specifier, self.current_module_path.as_deref()) {
-                Ok(p) => p,
-                Err(e) => {
-                    return self.create_rejected_promise(e);
-                }
-            };
-
-        // If we're NOT inside a static module load, load synchronously
-        if self.static_module_load_depth == 0 {
-            let module = match self.load_module(&resolved) {
-                Ok(m) => m,
-                Err(e) => {
-                    return self.create_rejected_promise(e);
-                }
-            };
-            let resolved_canon = resolved.canonicalize().unwrap_or(resolved.clone());
-            let mut stack = vec![];
-            if let Err(ref e) = self.inner_module_evaluation(&resolved_canon, &mut stack, 0) {
-                for m_path in &stack {
-                    if let Some(m) = self.module_registry.get(m_path) {
-                        let mut mb = m.borrow_mut();
-                        mb.evaluated = true;
-                        mb.is_evaluating = false;
-                        if mb.error.is_none() {
-                            mb.error = Some(e.clone());
-                        }
-                    }
-                }
-                return self.create_rejected_promise(e.clone());
-            }
-            // Drain microtasks to complete TLA module evaluation
-            self.drain_microtasks();
-            // Check if module evaluation failed (e.g. await rejected)
-            if let Some(err) = module.borrow().error.clone() {
-                return self.create_rejected_promise(err);
-            }
-            let ns = self.create_module_namespace(&module);
-            return self.create_resolved_promise(ns);
-        }
-
-        // Inside static module evaluation — defer to microtask so we don't
-        // preempt the current module's DFS evaluation order
-        let promise = self.create_promise_object();
-        let pid = if let JsValue::Object(ref o) = promise {
-            o.id
-        } else {
-            0
-        };
-        let resolved_path = resolved.clone();
-        let promise_root = promise.clone();
-
-        self.microtask_queue.push((
-            vec![promise_root],
-            Box::new(move |interp: &mut Interpreter| {
-                match interp.load_module(&resolved_path) {
-                    Ok(m) => {
-                        let resolved_canon = resolved_path
-                            .canonicalize()
-                            .unwrap_or(resolved_path.clone());
-                        let mut stack = vec![];
-                        if let Err(ref e) =
-                            interp.inner_module_evaluation(&resolved_canon, &mut stack, 0)
-                        {
-                            for m_path in &stack {
-                                if let Some(m) = interp.module_registry.get(m_path) {
-                                    let mut mb = m.borrow_mut();
-                                    mb.evaluated = true;
-                                    mb.is_evaluating = false;
-                                    if mb.error.is_none() {
-                                        mb.error = Some(e.clone());
-                                    }
-                                }
-                            }
-                            interp.reject_promise(pid, e.clone());
-                            return Completion::Normal(JsValue::Undefined);
-                        }
-                        interp.drain_microtasks();
-                        if let Some(err) = m.borrow().error.clone() {
-                            interp.reject_promise(pid, err);
-                            return Completion::Normal(JsValue::Undefined);
-                        }
-                        let ns = interp.create_module_namespace(&m);
-                        interp.fulfill_promise(pid, ns);
-                    }
-                    Err(e) => {
-                        interp.reject_promise(pid, e);
-                    }
-                }
-                Completion::Normal(JsValue::Undefined)
-            }),
-        ));
-
-        Completion::Normal(promise)
-    }
-
-    pub(crate) fn create_error_in_realm(
-        &mut self,
-        realm_id: usize,
-        error_type: &str,
-        msg: &str,
-    ) -> JsValue {
-        let old_realm = self.current_realm_id;
-        self.current_realm_id = realm_id;
-        let err = self.create_error(error_type, msg);
-        self.current_realm_id = old_realm;
-        err
-    }
-
-    pub(crate) fn get_wrapped_value(
-        &mut self,
-        caller_realm_id: usize,
-        value: &JsValue,
-    ) -> Result<JsValue, JsValue> {
-        match value {
-            JsValue::Undefined
-            | JsValue::Null
-            | JsValue::Boolean(_)
-            | JsValue::Number(_)
-            | JsValue::String(_)
-            | JsValue::Symbol(_)
-            | JsValue::BigInt(_) => Ok(value.clone()),
-            JsValue::Object(_) => {
-                if !self.is_callable(value) {
-                    return Err(self.create_error_in_realm(
-                        caller_realm_id,
-                        "TypeError",
-                        "ShadowRealm can only pass callable and primitive values across realm boundaries",
-                    ));
-                }
-                self.wrapped_function_create(caller_realm_id, value)
-            }
-        }
-    }
-
-    pub(crate) fn wrapped_function_create(
-        &mut self,
-        caller_realm_id: usize,
-        target_func: &JsValue,
-    ) -> Result<JsValue, JsValue> {
-        let old_realm = self.current_realm_id;
-        self.current_realm_id = caller_realm_id;
-
-        let func_obj = self.create_object();
-        let func_id = func_obj.borrow().id.unwrap();
-        {
-            let mut o = func_obj.borrow_mut();
-            o.prototype = self.realms[caller_realm_id].function_prototype.clone();
-            o.class_name = "Function".to_string();
-            o.callable = Some(JsFunction::native("".to_string(), 0, |_, _, _| {
-                Completion::Normal(JsValue::Undefined)
-            }));
-            if let JsValue::Object(tf) = target_func {
-                o.wrapped_target_function_id = Some(tf.id);
-            }
-            o.wrapped_caller_realm_id = Some(caller_realm_id);
-        }
-        self.function_realm_map.insert(func_id, caller_realm_id);
-
-        self.current_realm_id = old_realm;
-
-        // CopyNameAndLength (spec §10.4.2.4) — any error becomes TypeError from callerRealm
-        let length_val = if let JsValue::Object(tf) = target_func {
-            // HasOwnProperty via [[GetOwnProperty]] (invokes proxy trap if proxy)
-            let has_own_length = match self.proxy_get_own_property_descriptor(tf.id, "length") {
-                Ok(JsValue::Undefined) => false,
-                Ok(_) => true,
-                Err(_) => {
-                    return Err(self.create_error_in_realm(
-                        caller_realm_id,
-                        "TypeError",
-                        "WrappedFunctionCreate: error getting length descriptor",
-                    ));
-                }
-            };
-            if has_own_length {
-                match self.get_object_property(tf.id, "length", target_func) {
-                    Completion::Normal(v) => v,
-                    Completion::Throw(_) => {
-                        return Err(self.create_error_in_realm(
-                            caller_realm_id,
-                            "TypeError",
-                            "WrappedFunctionCreate: error getting length",
-                        ));
-                    }
-                    _ => JsValue::Number(0.0),
-                }
-            } else {
-                JsValue::Number(0.0)
-            }
-        } else {
-            JsValue::Number(0.0)
-        };
-
-        let computed_length = match length_val {
-            JsValue::Number(n) => {
-                if n == f64::INFINITY {
-                    f64::INFINITY
-                } else if n == f64::NEG_INFINITY || n < 0.0 {
-                    0.0
-                } else {
-                    n.trunc().max(0.0)
-                }
-            }
-            _ => 0.0,
-        };
-
-        let name_str = if let JsValue::Object(tf) = target_func {
-            match self.get_object_property(tf.id, "name", target_func) {
-                Completion::Normal(JsValue::String(s)) => s.to_string(),
-                Completion::Normal(_) => String::new(),
-                Completion::Throw(_) => {
-                    return Err(self.create_error_in_realm(
-                        caller_realm_id,
-                        "TypeError",
-                        "WrappedFunctionCreate: error getting name",
-                    ));
-                }
-                _ => String::new(),
-            }
-        } else {
-            String::new()
-        };
-
-        if let Some(obj) = self.get_object(func_id) {
-            obj.borrow_mut().insert_property(
-                "length".to_string(),
-                PropertyDescriptor::data(JsValue::Number(computed_length), false, false, true),
-            );
-            obj.borrow_mut().insert_property(
-                "name".to_string(),
-                PropertyDescriptor::data(
-                    JsValue::String(crate::types::JsString::from_str(&name_str)),
-                    false,
-                    false,
-                    true,
-                ),
-            );
-        }
-
-        Ok(JsValue::Object(crate::types::JsObject { id: func_id }))
-    }
-
-    pub(crate) fn call_wrapped_function(
-        &mut self,
-        wrapper_id: u64,
-        _this_val: &JsValue,
-        args: &[JsValue],
-    ) -> Completion {
-        let (target, caller_realm_id) = {
-            let obj = match self.get_object(wrapper_id) {
-                Some(o) => o,
-                None => {
-                    return Completion::Throw(
-                        self.create_type_error("WrappedFunction: missing target"),
-                    );
-                }
-            };
-            let b = obj.borrow();
-            let target_id = match b.wrapped_target_function_id {
-                Some(id) => id,
-                None => {
-                    return Completion::Throw(
-                        self.create_type_error("WrappedFunction: missing target"),
-                    );
-                }
-            };
-            let caller_realm_id = b.wrapped_caller_realm_id.unwrap_or(self.current_realm_id);
-            (
-                JsValue::Object(crate::types::JsObject { id: target_id }),
-                caller_realm_id,
-            )
-        };
-
-        let target_realm_id = match self.get_function_realm(&target) {
-            Ok(r) => r,
-            Err(e) => return Completion::Throw(e),
-        };
-
-        // Wrap arguments into target realm
-        let mut wrapped_args = Vec::with_capacity(args.len());
-        for arg in args {
-            match self.get_wrapped_value(target_realm_id, arg) {
-                Ok(v) => wrapped_args.push(v),
-                Err(_) => {
-                    return Completion::Throw(self.create_error_in_realm(
-                        caller_realm_id,
-                        "TypeError",
-                        "WrappedFunction: argument is not a primitive or callable",
-                    ));
-                }
-            }
-        }
-
-        // Call target in its realm
-        let old_realm = self.current_realm_id;
-        self.current_realm_id = target_realm_id;
-        let result = self.call_function(&target, &JsValue::Undefined, &wrapped_args);
-        self.current_realm_id = old_realm;
-
-        let result_val = match result {
-            Completion::Normal(v) => v,
-            Completion::Empty => JsValue::Undefined,
-            _ => {
-                return Completion::Throw(self.create_error_in_realm(
-                    caller_realm_id,
-                    "TypeError",
-                    "WrappedFunction: error in target function",
-                ));
-            }
-        };
-
-        match self.get_wrapped_value(caller_realm_id, &result_val) {
-            Ok(v) => Completion::Normal(v),
-            Err(e) => Completion::Throw(e),
-        }
-    }
-
-    pub(crate) fn perform_realm_eval(
-        &mut self,
-        source_text: &str,
-        caller_realm_id: usize,
-        eval_realm_id: usize,
-    ) -> Completion {
-        use crate::parser::Parser;
-
-        let program = {
-            let mut parser = match Parser::new(source_text) {
-                Ok(p) => p,
-                Err(_) => {
-                    return Completion::Throw(self.create_error_in_realm(
-                        caller_realm_id,
-                        "SyntaxError",
-                        "Invalid source text",
-                    ));
-                }
-            };
-            match parser.parse_program() {
-                Ok(prog) => prog,
-                Err(e) => {
-                    return Completion::Throw(self.create_error_in_realm(
-                        caller_realm_id,
-                        "SyntaxError",
-                        &e.to_string(),
-                    ));
-                }
-            }
-        };
-
-        let old_realm = self.current_realm_id;
-        self.current_realm_id = eval_realm_id;
-        let is_strict = program.body_is_strict;
-        // Per spec §B.3.6.2 PerformShadowRealmEval:
-        // Strict: both var and lex are new strict envs (isolates everything)
-        // Non-strict: var goes to global, lex is fresh child of global (isolates let/const)
-        let global = self.realm().global_env.clone();
-        let (var_env, lex_env) = if is_strict {
-            let new_env = Environment::new_function_scope(Some(global));
-            new_env.borrow_mut().strict = true;
-            (new_env.clone(), new_env)
-        } else {
-            let lex_env = Environment::new(Some(global.clone()));
-            (global, lex_env)
-        };
-        // Hoist var/function declarations to var_env
-        if let Err(e) = self.eval_declaration_instantiation(
-            &program.body,
-            &var_env,
-            &lex_env,
-            is_strict,
-            false,
-            &lex_env,
-        ) {
-            self.current_realm_id = old_realm;
-            return Completion::Throw(e);
-        }
-        // Execute body in lex_env
-        self.call_stack_envs.push(lex_env.clone());
-        let mut result = Completion::Empty;
-        for stmt in &program.body {
-            self.maybe_gc();
-            match self.exec_statement(stmt, &lex_env) {
-                Completion::Normal(v) => result = Completion::Normal(v),
-                Completion::Empty => {}
-                other => {
-                    result = other;
-                    break;
-                }
-            }
-        }
-        self.call_stack_envs.pop();
-        self.drain_microtasks();
-        self.current_realm_id = old_realm;
-
-        let result_val = match result {
-            Completion::Normal(v) => v,
-            Completion::Empty => JsValue::Undefined,
-            _ => {
-                return Completion::Throw(self.create_error_in_realm(
-                    caller_realm_id,
-                    "TypeError",
-                    "ShadowRealm evaluate error",
-                ));
-            }
-        };
-
-        match self.get_wrapped_value(caller_realm_id, &result_val) {
-            Ok(v) => Completion::Normal(v),
-            Err(e) => Completion::Throw(e),
         }
     }
 }
